requirements.txt
django==4.2.29
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
requests==2.32.3
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
